# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestão Jurídica",
        "description": "Sistema de Gestão Jurídica",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tipopedidos": {
            "get": {
                "tags": [
                    "Tipopedidos"
                ],
                "summary": "Listar Tipopedidos",
                "operationId": "tipopedido/index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order  accepts 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tipopedido": {
            "post": {
                "tags": [
                    "Tipopedidos"
                ],
                "summary": "Cadastrar Tipopedido",
                "operationId": "tipopedido/store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "aluno_id",
                                    "materia_id",
                                    "periodo_id",
                                    "tipopedido",
                                    "aprovado"
                                ],
                                "properties": {
                                    "aluno_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "materia_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "periodo_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "tipopedido": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "9"
                                    },
                                    "aprovado": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sim"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tipopedido/{id}": {
            "get": {
                "tags": [
                    "Tipopedidos"
                ],
                "summary": "Pesquisar Tipopedido",
                "operationId": "tipopedido/show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Tipopedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipopedidos"
                ],
                "summary": "Editar Tipopedido",
                "operationId": "tipopedido/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Tipopedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "aluno_id",
                                    "materia_id",
                                    "periodo_id",
                                    "tipopedido",
                                    "aprovado"
                                ],
                                "properties": {
                                    "aluno_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "materia_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "periodo_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "tipopedido": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "9"
                                    },
                                    "aprovado": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sim"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipopedidos"
                ],
                "summary": "Excluir Tipopedido",
                "operationId": "tipopedido/destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Tipopedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}